removed cellEvent overloads; moved dataModel wrappers from Local to Behavior
