+++ src/behaviors/rowProperties.js
@@ -46,7 +46,7 @@ exports.mixin = {
      * @param {boolean} [prototype] - Prototype for a new properties object when one does not already exist. If you don't define this and one does not already exist, this call will return `undefined`.
      * Typical defined value is `null`, which creates a plain object with no prototype, or `Object.prototype` for a more "natural" object.
      * _(Required when 3rd param provided.)_
-     * @param {dataModelAPI} [dataModel=this.dataModel] - This is the subgrid. You only need to provide the subgrid when it is not the data subgrid _and_ you did not give a `CellEvent` object in the first param (which already knows what subgrid it's in).
+     * @param {DataModel} [dataModel=this.dataModel] - This is the subgrid. You only need to provide the subgrid when it is not the data subgrid _and_ you did not give a `CellEvent` object in the first param (which already knows what subgrid it's in).
      * @returns {object|undefined} The row properties object which will be one of:
      * * object - existing row properties object or new row properties object created from `prototype`; else
      * * `false` - row found but no existing row properties object and `prototype` was not defined; else
@@ -67,7 +67,7 @@ exports.mixin = {
      * @memberOf Behavior#
      * @param {number|CellEvent} yOrCellEvent - Data row index local to `dataModel`; or a `CellEvent` object.
      * @param {object|undefined} properties - The new row properties object. If `undefined`, this call is a no-op.
-     * @param {dataModelAPI} [dataModel=this.dataModel] - This is the subgrid. You only need to provide the subgrid when it is not the data subgrid _and_ you did not give a `CellEvent` object in the first param (which already knows what subgrid it's in).
+     * @param {DataModel} [dataModel=this.dataModel] - This is the subgrid. You only need to provide the subgrid when it is not the data subgrid _and_ you did not give a `CellEvent` object in the first param (which already knows what subgrid it's in).
      */
     setRowProperties: function(yOrCellEvent, properties, dataModel) {
         if (!properties) {
@@ -93,7 +93,7 @@ exports.mixin = {
      * @param {number|CellEvent} yOrCellEvent - Data row index local to `dataModel`; or a `CellEvent` object.
      * @param {string} key - The property name.
      * @param value - The new property value.
-     * @param {dataModelAPI} [dataModel=this.dataModel] - This is the subgrid. You only need to provide the subgrid when it is not the data subgrid _and_ you did not give a `CellEvent` object in the first param (which already knows what subgrid it's in).
+     * @param {DataModel} [dataModel=this.dataModel] - This is the subgrid. You only need to provide the subgrid when it is not the data subgrid _and_ you did not give a `CellEvent` object in the first param (which already knows what subgrid it's in).
      */
     setRowProperty: function(yOrCellEvent, key, value, dataModel) {
         var rowProps;
@@ -122,7 +122,7 @@ exports.mixin = {
      * @memberOf Behavior#
      * @param {number|CellEvent} yOrCellEvent - Data row index local to `dataModel`; or a `CellEvent` object.
      * @param {object|undefined} properties - An object containing new property values(s) to assign to the row properties. If `undefined`, this call is a no-op.
-     * @param {dataModelAPI} [dataModel=this.dataModel] - This is the subgrid. You only need to provide the subgrid when it is not the data subgrid _and_ you did not give a `CellEvent` object in the first param (which already knows what subgrid it's in).
+     * @param {DataModel} [dataModel=this.dataModel] - This is the subgrid. You only need to provide the subgrid when it is not the data subgrid _and_ you did not give a `CellEvent` object in the first param (which already knows what subgrid it's in).
      */
     addRowProperties: function(yOrCellEvent, properties, dataModel, rowProps) {
         if (!properties) {
@@ -156,7 +156,7 @@ exports.mixin = {
     /**
      * @memberOf Behavior#
      * @param {number} yOrCellEvent - Data row index local to `dataModel`.
-     * @param {dataModelAPI} [dataModel=this.dataModel]
+     * @param {DataModel} [dataModel=this.dataModel]
      * @returns {number} The row height in pixels.
      */
     getRowHeight: function(yOrCellEvent, dataModel) {
@@ -169,7 +169,7 @@ exports.mixin = {
      * @desc set the pixel height of a specific row
      * @param {number} yOrCellEvent - Data row index local to dataModel.
      * @param {number} height - pixel height
-     * @param {dataModelAPI} [dataModel=this.dataModel]
+     * @param {DataModel} [dataModel=this.dataModel]
      */
     setRowHeight: function(yOrCellEvent, height, dataModel) {
         this.setRowProperty(yOrCellEvent, 'height', height, dataModel);
